Reorder messages as in draft-10
