+++ session_helpers.go
@@ -1,0 +1,62 @@
+package moqtransport
+
+import (
+	"github.com/mengelbart/moqtransport/internal/wire"
+	"github.com/quic-go/quic-go/quicvarint"
+)
+
+func compileMessage(msg wire.ControlMessage) []byte {
+	buf := make([]byte, 16, 1500)
+	buf = append(buf, msg.Append(buf[16:])...)
+	length := len(buf[16:])
+
+	typeLenBuf := quicvarint.Append(buf[:0], uint64(msg.Type()))
+	typeLenBuf = quicvarint.Append(typeLenBuf, uint64(length))
+
+	n := copy(buf[0:16], typeLenBuf)
+	buf = append(buf[:n], buf[16:]...)
+
+	return buf
+}
+
+func validatePathParameter(setupParameters wire.Parameters, protocolIsQUIC bool) (string, error) {
+	pathParam, ok := setupParameters[wire.PathParameterKey]
+	if !ok {
+		if !protocolIsQUIC {
+			return "", nil
+		}
+		return "", errMissingPathParameter
+	}
+	if !protocolIsQUIC {
+		return "", errUnexpectedPathParameter
+	}
+	pathParamValue, ok := pathParam.(wire.StringParameter)
+	if !ok {
+		return "", errInvalidPathParameterType
+	}
+	return pathParamValue.Value, nil
+}
+
+func validateMaxSubscribeIDParameter(setupParameters wire.Parameters) (uint64, error) {
+	maxSubscribeIDParam, ok := setupParameters[wire.MaxSubscribeIDParameterKey]
+	if !ok {
+		return 0, nil
+	}
+	maxSubscribeIDParamValue, ok := maxSubscribeIDParam.(wire.VarintParameter)
+	if !ok {
+		return 0, errInvalidMaxSubscribeIDParameterType
+	}
+	return maxSubscribeIDParamValue.Value, nil
+}
+
+func validateAuthParameter(subscribeParameters wire.Parameters) (string, error) {
+	authParam, ok := subscribeParameters[wire.AuthorizationParameterKey]
+	if !ok {
+		return "", nil
+	}
+	authParamValue, ok := authParam.(wire.StringParameter)
+	if !ok {
+		return "", errInvalidAuthParameterType
+	}
+	return authParamValue.Value, nil
+}
